requirements.txt
flask
flask-pymongo
mongoengine
PyPDF2
python-docx
spacy
nltk
scikit-learn
